requirements.txt
# Consolidated requirements for the entire project (interface + simulator + map module)
# Notes:
# - Standard library modules are NOT listed here
# - Version pins aim to be broadly compatible across modules

# Web interface
streamlit>=1.28.0

# Core scientific stack
numpy>=1.19.5
pandas>=2.0.0
scipy>=1.9.0

# Imaging
Pillow>=8.1.0

# Plotting (simulator)
matplotlib>=3.3.4
cycler>=0.10.0
kiwisolver>=1.3.1
pyparsing>=2.4.7
python-dateutil>=2.8.1
six>=1.15.0
colour==0.1.5

# Optimization (NSGA-II and related via pymoo)
pymoo>=0.6.0
